refactor(editor): Move versions.store into @n8n/stores
Relocates `versions.store` (313 LOC) from `editor-ui` into `@n8n/stores`,
continuing the frontend modularization wave. The `@/app/stores/versions.store`
path becomes a re-export shim, so all 12 existing importers are untouched.

Now possible without a package cycle because N8N-100 sank `@n8n/composables`
below `@n8n/stores`: the store imports `useToast`, `useTelemetry` and
`useStorage` directly from `@n8n/composables`.

- `NotificationHandle` resolves from `@n8n/composables/useToast` (which declares
  it locally) rather than `element-plus`, so no new dependency is added to
  `@n8n/stores`.
- The four constants the store owns move per-symbol into
  `versions.constants.ts`. The two `LOCAL_STORAGE_*` keys had no other consumer;
  the two modal keys are re-exported from `@/app/constants/modals` for app-side
  modal registration.
- Test mocks repointed, and the `as any` users-store stubs replaced with a typed
  `mock<T>()` helper (the stores package lints `no-unsafe-argument` as an error).
- The `vue-router` mock is dropped: it existed only to serve the real `useToast`,
  which the suite factory-mocks.
